Geometry Generation

Documentation for functions used to generate potential geometry during the population growth in the genetic algorithm optimization.

geometry_configuration.load_JSON — Method

load_JSON()

load the configuration and geometry json files. note that the radius, length, width, and thickness must be followed by the value set for "scaling_factor" in the configuration.json file all keys must be present for all shapes

Expected Formatting for 2 shape geometry input
[
                {
                    "shape": "Rod",
                    "radius": 100e-9,
                    "length": 100e-9,
                    "width": 100e-9,
                    "thickness": 100e-9,
                    "material": "gold",
                    "position": [201.4, 101.1, 249.1]
                },
                {
                    "shape": "Rod",
                    "radius": 100e-9,
                    "length": 100e-9,
                    "width": 100e-9,
                    "thickness": 100e-9,
                    "material": "gold",
                    "position": [201.4, 101.1, 249.1]
                }
]

Geometry Keys:
    "shape": Expected values: "Rod", "Rectangle", or "Triangle"
    "radius": radius of rod in the XZ-plane
        can be any positive floating point number
    "length": number of units material spans along the Z-axis
        can be any positive floating point number
    "width": number of units material spans along the X-axis
        can be any positive floating point number
    "Thickness": number of units material spans along the Y-axis
        can be any positive floating point number
    "material": defines the color of the shape in the visualizer
        currently the only defined colors are blue for "air"
        and red for any other value
    "position": coordinateds for the center of the shape in
        a three value ordered pair, can be any positive float

Expected Formatting for 2 shape geometry input
            [
                {
                    "scaling_factor": 1E9,
                    "x_span": 450,
                    "y_span": 300,
                    "z_span": 900
                }
            ]

Configuration Keys:
    "scaling_factor": corresponds to the units of the geometry value
    "x_span": sets the maximum value of the slider for the x-coordinate
    "y_span": sets the maximum value of the slider for the y-coordinate
    "z_span": sets the maximum value of the slider for the z-coordinate

geometry_configuration.get_vertices — Method

get_vertices(x1_coordinate_slider, x2_coordinate_slider, plane, L, W, T, R, shape)

geometry_configuration.set_plane — Method

set_plane(plane, L, W, T)

geometry_configuration.rectangle_vertices — Method

rectangle_vertices(x1_coordinate_slider, x2_coordinate_slider, plane, L, W, T)

geometry_configuration.triangle_vertices — Method

triangle_vertices(x1_coordinate_slider, x2_coordinate_slider, W, L)

geometry_configuration.circle_vertices — Method

circle_vertices(x1_coordinate_slider, x2_coordinate_slider, R)

geometry_configuration.place_shape — Method

place_shape(x1_coordinate_slider, x2_coordinate_slider, plane, L, W, T, R, shape)